requirements.txt
pygame>=2.5.0
streamlit>=1.28.0
Pillow>=9.0.0

